style: move export panel to sidebar
