+++ web-editor/src/component/google-maps/GoogleMapsSettings.jsx
@@ -18,6 +18,7 @@ import { IndicatorSetting } from "../IndicatorSetting";
 import PullKeyInput from "../PullKeyInput";
 import { ZoomSlider } from "../ZoomSlider";
 import { GoogleAPIKeyDialog } from "./GoogleAPIKeyDialog";
+import OverlayExportPanel from "../OverlayExportPanel";
 
 function isValidJSON(string) {
   if (!string) {
@@ -48,6 +49,19 @@ export const GoogleMapsSettings = ({
   const [showStyleDialog, setShowStyleDialog] = React.useState(false);
   const [showGoogleAPIKeyDialog, setShowGoogleAPIKeyDialog] =
     React.useState(false);
+  const jsonMapStyle = mapStyle.valid ? JSON.parse(mapStyle.value) : {};
+  const styleB64 = encodeURIComponent(
+    window.btoa(JSON.stringify(jsonMapStyle))
+  );
+  const indicatorStyleB64 = encodeURIComponent(
+    window.btoa(JSON.stringify(indicatorStyle))
+  );
+
+  const url = `https://overlays.rtirl.com/googlemaps.html?key=${
+    pullKey.value
+  }&api_key=${apiKey}&style=${styleB64}&zoom=${zoom}${
+    fullscreen ? "&fullscreen=1" : ""
+  }&indicatorStyle=${indicatorStyleB64}`;
 
   return (
     <Box
@@ -64,10 +78,15 @@ export const GoogleMapsSettings = ({
         spacing={1}
         textAlign="left"
       >
-
         <PullKeyInput pullKey={pullKey} onKeyChange={onPullKeyChange} />
 
-        <Box bgcolor="black" component="form" noValidate autoComplete="off"  sx={{ paddingTop: "8px", paddingLeft: "24px", paddingBottom: "8px" }} >
+        <Box
+          bgcolor="black"
+          component="form"
+          noValidate
+          autoComplete="off"
+          sx={{ paddingTop: "8px", paddingLeft: "24px", paddingBottom: "8px" }}
+        >
           <TextField
             fullWidth
             required
@@ -99,6 +118,18 @@ export const GoogleMapsSettings = ({
           />
         </Box>
 
+        <Box bgcolor="black" sx={{ marginTop: "12px" }}>
+          <Typography sx={{ paddingLeft: "24px", paddingTop: "10px" }}>
+            Export
+          </Typography>
+          <OverlayExportPanel
+            overlayDescription="Goole Maps Overlay URL"
+            isExportable={pullKey.valid && apiKey && mapStyle.valid}
+            url={url}
+            type="google_maps_overlay"
+          />
+        </Box>
+
         {/* Style Id form */}
         <Box
           bgcolor="black"
@@ -109,7 +140,7 @@ export const GoogleMapsSettings = ({
           onSubmit={(e) => {
             e.preventDefault();
           }}
-          sx={{paddingTop:"15px", paddingBottom:"5px"}}
+          sx={{ paddingTop: "15px", paddingBottom: "5px" }}
         >
           <TextField
             fullWidth
@@ -164,14 +195,18 @@ export const GoogleMapsSettings = ({
               <Checkbox
                 value={fullscreen}
                 onChange={(event) => setFullscreen(event.target.checked)}
-                sx={{paddingLeft: "24px", paddingTop: "5px", paddingBottom:"5px"}}
+                sx={{
+                  paddingLeft: "24px",
+                  paddingTop: "5px",
+                  paddingBottom: "5px",
+                }}
               />
             }
             label="Fullscreen"
           />
         </Box>
 
-        <Box bgcolor="black" sx={{padding:"10px"}}>
+        <Box bgcolor="black" sx={{ padding: "10px" }}>
           <IndicatorSetting
             indicatorStyle={indicatorStyle}
             setIndicatorStyle={setIndicatorStyle}
